Reorganiza arquivos de componentes
